Adds cat fact for todo placeholder
I wanted to have some whimsy with the next todo input.  I had the idea
that we could retrieve a funny cat fact to display when the user hadn't
put anything in the text box yet.

This commit calls out to the cat fact API to retrieve a list of cat
facts and then chooses a random one from the list.  It also changes the
cat fact each time the user adds a new todo item.